requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
python-dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.12.0
lxml>=4.9.2
aiohttp>=3.8.4
supabase>=1.0.3
tqdm>=4.65.0
python-multipart>=0.0.6
selenium>=4.8.0
undetected-chromedriver>=3.5.0
webdriver-manager>=3.8.5
pandas>=1.5.3
gunicorn>=20.1.0
